Split PortableThreadPool.WorkerThread start and loop body
For browser-wasm we will need to start the worker thread in a special
way, and use callbacks to run the loop body.

Current PR is just refactoring existing code. No functional change.
